Fix thread safety, GC pressure, and lifecycle issues in SudoVDA features
- Make `connected` volatile for safe cross-thread reads during reconnect
- Make `serverStatsListener` volatile for JNI callback thread safety
- Remove log spam from no-op sendWifiQuality stub
- Fix RTT cast truncation in MediaCodecDecoderRenderer
- Reuse StringBuilder in StatsOverlay/ReconnectOverlay onDraw to reduce GC
- Clear pending handler callbacks in WifiMonitor.stop()

+++ app/src/main/java/com/limelight/binding/video/MediaCodecDecoderRenderer.java
@@ -1883,7 +1883,7 @@ public class MediaCodecDecoderRenderer extends VideoDecoderRenderer implements C
                 // Also send structured stats for the SudoVDA stats overlay
                 float totalTimeMs = lastTwo.totalFrames > 0 ?
                         (float)lastTwo.totalTimeMs / lastTwo.totalFrames : 0;
-                float networkLatencyMs = (int)(rttInfo >> 32); // average RTT
+                float networkLatencyMs = (float)(rttInfo >> 32); // average RTT
                 String codecName;
                 if ((videoFormat & MoonBridge.VIDEO_FORMAT_MASK_H264) != 0) {
                     codecName = "H.264";
